+++ BlazorDeviceControl/Shared/Section/SectionLogs.razor.cs
@@ -53,7 +53,7 @@ namespace BlazorDeviceControl.Shared.Section
                     if (AppSettings.DataAccess != null)
                     {
                         object[] objects = AppSettings.DataAccess.Crud.GetEntitiesNativeObject(
-                            SqlQueries.DbServiceManaging.Tables.Logs.GetLogs);
+                            SqlQueries.DbServiceManaging.Tables.Logs.GetLogs(IsShowTop100 ? 100 : 0));
                         Items = new List<LogQuickEntity>().ToList<BaseEntity>();
                         foreach (object obj in objects)
                         {