+++ WeightCore/Gui/GuiUtils.cs
@@ -2,6 +2,7 @@
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
 using DataCore;
+using System;
 using System.Diagnostics;
 using System.Windows;
 using System.Windows.Forms;
@@ -43,6 +44,23 @@ namespace WeightCore.Gui
             else
                 Process.Start(isDebug
                     ? "https://device-control-dev.kolbasa-vs.local/" : "https://device-control.kolbasa-vs.local/");
+        }
+
+        public static void ShowWpfCatch(IWin32Window owner, string message, int lineNumber, string memberName)
+        {
+            using WpfPageLoader wpfPageLoader = new(ProjectsEnums.Page.MessageBox, false) { Width = 700, Height = 400 };
+            wpfPageLoader.MessageBox.Caption = LocalizationData.ScalesUI.Exception;
+            wpfPageLoader.MessageBox.Message =
+                @$"{@LocalizationData.ScalesUI.Method}: {memberName}." + Environment.NewLine +
+                $"{@LocalizationData.ScalesUI.Line}: {lineNumber}." + Environment.NewLine + Environment.NewLine + message;
+            wpfPageLoader.MessageBox.ButtonOkVisibility = System.Windows.Visibility.Visible;
+            wpfPageLoader.MessageBox.Localization();
+            if (owner != null)
+                wpfPageLoader.ShowDialog(owner);
+            else
+                wpfPageLoader.ShowDialog();
+            wpfPageLoader.Close();
+            wpfPageLoader.Dispose();
         }
 
         #endregion